@@ -1,38 +1,38 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Version>3.0.0</Version>
-    <FileVersion>3.0.0</FileVersion>
-    <AssemblyVersion>3.0.0</AssemblyVersion>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net9.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Version>3.0.0</Version>
+		<FileVersion>3.0.0</FileVersion>
+		<AssemblyVersion>3.0.0</AssemblyVersion>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="AspNetCore.Identity.Mongo" Version="9.0.0" />
-    <PackageReference Include="Cadmus.Api.Controllers" Version="9.0.10" />
-    <PackageReference Include="Cadmus.Api.Models" Version="9.0.10" />
-    <PackageReference Include="Cadmus.Api.Services" Version="9.0.10" />
-    <PackageReference Include="Cadmus.Graph.Extras" Version="7.0.4" />
-    <PackageReference Include="Cadmus.Seed.Epigraphy.Parts" Version="4.0.1" />
-    <PackageReference Include="Cadmus.Seed.General.Parts" Version="6.1.3" />
-    <PackageReference Include="Cadmus.Seed.Geo.Parts" Version="3.0.5" />
-    <PackageReference Include="Cadmus.Seed.Philology.Parts" Version="8.2.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.10" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="8.0.10" />
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.10" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="Polly" Version="8.4.2" />
-    <PackageReference Include="Serilog" Version="4.1.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="8.0.3" />
-    <PackageReference Include="Serilog.Exceptions" Version="8.4.0" />
-    <PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
-    <PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
-    <PackageReference Include="Serilog.Sinks.MongoDB" Version="6.0.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.9.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Cadmus.Api.Config" Version="10.1.1" />
+		<PackageReference Include="Cadmus.Api.Controllers" Version="10.1.0" />
+		<PackageReference Include="Cadmus.Api.Models" Version="10.1.0" />
+		<PackageReference Include="Cadmus.Api.Services" Version="10.1.0" />
+		<PackageReference Include="Cadmus.Graph.Extras" Version="8.0.0" />
+		<PackageReference Include="Cadmus.Seed.Epigraphy.Parts" Version="5.0.0" />
+		<PackageReference Include="Cadmus.Seed.General.Parts" Version="7.0.0" />
+		<PackageReference Include="Cadmus.Seed.Geo.Parts" Version="3.0.5" />
+		<PackageReference Include="Cadmus.Seed.Philology.Parts" Version="9.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="9.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+		<PackageReference Include="Polly" Version="8.5.0" />
+		<PackageReference Include="Scalar.AspNetCore" Version="1.2.39" />
+		<PackageReference Include="Serilog" Version="4.1.0" />
+		<PackageReference Include="Serilog.AspNetCore" Version="8.0.3" />
+		<PackageReference Include="Serilog.Exceptions" Version="8.4.0" />
+		<PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
+		<PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
+		<PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
+		<PackageReference Include="Serilog.Sinks.MongoDB" Version="7.0.0" />
+	</ItemGroup>
 
 </Project>
